refactor: repackage logger
This was previously in a utils module that was not being used for other
tools. It was also missing an __init__.py and was not being properly
exported. This pattern is slightly less verbose.
