+++ hm_pyhelper/tests/test_logger.py
@@ -1,11 +1,11 @@
 from unittest import TestCase
-from hm_pyhelper.utils.logger import get_logger, _log_format
+from hm_pyhelper.logger import get_logger, _log_format
 import re
 import logging
 
 
-class TestExample(TestCase):
-    def test_logging(self):
+class TestLogger(TestCase):
+    def test_get_logger(self):
         logger = get_logger(__name__)
 
         with self.assertLogs() as captured:
@@ -20,7 +20,8 @@ class TestExample(TestCase):
         # Do not check timestamp and filepath because those change
         # based on the environment and run time
         expected_partial_output_regex = re.escape(
-            " - [DEBUG] - test_logger - (test_logger.py).test_logging -- ")
+            " - [DEBUG] - hm_pyhelper.tests.test_logger -" +
+            " (test_logger.py).test_get_logger -- ")
         expected_output_regex = ".*" + \
             expected_partial_output_regex + ".*" + \
             " - Hello world."
